refactor: Extract this action into her own component
