+++ src/modules/views/Drive/DriveFolderView.jsx
@@ -21,12 +21,12 @@ import {
   download,
   trash,
   rename,
-  move,
   duplicate,
   qualify,
   versions,
   hr
 } from 'modules/actions'
+import { moveTo } from 'modules/actions/components/moveTo'
 import { makeExtraColumnsNamesFromMedia } from 'modules/certifications'
 import { useExtraColumns } from 'modules/certifications/useExtraColumns'
 import AddMenuProvider from 'modules/drive/AddMenu/AddMenuProvider'
@@ -145,7 +145,7 @@ const DriveFolderView = () => {
       hr,
       qualify,
       rename,
-      move,
+      moveTo,
       duplicate,
       hr,
       versions,
